requirements.txt
# Customer Support Agent - Requirements
# Python 3.8+

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: For enhanced AI capabilities (if you want to use OpenAI or other LLMs)
# openai>=1.0.0
# langchain>=0.1.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

